% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database_management.R
\name{checkAnnotatedTableFormat}
\alias{checkAnnotatedTableFormat}
\title{checkAnnotatedTableFormat}
\usage{
checkAnnotatedTableFormat(annotated_table)
}
\arguments{
\item{annotated_table}{peak table with annotation column as produced by detectHomologues.}
}
\value{
TRUE if annotated table matches assertions. Stops program if not.
}
\description{
checkAnnotatedTableFormat
}
\keyword{internal}
