% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sdbCheckContained.R
\name{checkAnyMatch}
\alias{checkAnyMatch}
\title{checkAnyMatch}
\usage{
checkAnyMatch(
  target_mz,
  target_rt,
  homologue_id,
  ppm_tolerance,
  rt_tolerance,
  peak_table
)
}
\arguments{
\item{target_mz}{Target mass to charge ratio.}

\item{target_rt}{Target retention time.}

\item{homologue_id}{Current peak homologue id.}

\item{ppm_tolerance}{Tolerance in part per million.}

\item{rt_tolerance}{Retention time tolerance. Beware of peak table retention time units!}

\item{peak_table}{A peak table (tibble) with mz and rt columns at least.}
}
\value{
NA or the matched homologue_id
}
\description{
Helper function checks whether there is any match within ppm and rt tolerance of some target mz and rt tuple within the target df. Target df needs to contain two columns: mz, rt. Function returns NA if no match, and the homologue_id corresponding to the mz-rt tuple if there is one or more matches.
}
\keyword{internal}
