% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{checkDetectHomologueSearchParameters}
\alias{checkDetectHomologueSearchParameters}
\title{checkDetectHomologueSearchParameters}
\usage{
checkDetectHomologueSearchParameters(
  mz_min,
  mz_max,
  rt_min,
  rt_max,
  ppm_tolerance,
  search_mode,
  mz_steps,
  min_series_length,
  step_mode
)
}
\arguments{
\item{mz_min}{Minimum mz increment for a new peak to be considered (untargeted, absolute value). Must be 0 or larger. Needs to be set for untargeted search mode.}

\item{mz_max}{Maximum mz increment for a new peak to be considered (untargeted, absolute value). Must be mz_min or larger. Needs to be set for untargeted search mode.}

\item{rt_min}{Minimum rt increment for a new peak to be considered. Must be 0 or larger.}

\item{rt_max}{Maximum rt increment for a new peak to be considered. Must be rt_min or larger.}

\item{ppm_tolerance}{Error tolerance in ppm (error relative to last chain node mz + mz step-size).}

\item{search_mode}{"targeted" or "untargeted" determining the search approach.}

\item{mz_steps}{Vector of mz step sizes to be searched around. Needs to be set for targeted search mode.}

\item{min_series_length}{Minimum size of a repetitive series for it to be considered a homologue. The minimum plausible value is 3. Any value below 5 is likely to return spurious results unless ppm constrains are highly rigid.}

\item{step_mode}{"increment" or "decrement" to indicate whether mz steps are to be looked for in an increasing or decreasing mass-to-charge ratio value over retention time fashion.}
}
\value{
Boolean TRUE if all settings in valid ranges.
}
\description{
Helper function that does rough checks for input parameter validity.
}
\keyword{internal}
