% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{checkPeakTableFormat}
\alias{checkPeakTableFormat}
\title{checkPeakTableFormat}
\usage{
checkPeakTableFormat(peak_table)
}
\arguments{
\item{peak_table}{}
}
\value{
Boolean True if peak table has the right format.
}
\description{
Helper function that checks whether the provided peak table
}
\keyword{internal}
