% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_database_management.R
\name{checkSeriesDBFormat}
\alias{checkSeriesDBFormat}
\title{checkSeriesDBFormat}
\usage{
checkSeriesDBFormat(seriesdb)
}
\arguments{
\item{seriesdb}{seriesdb object as produced by sdbCreate()}
}
\value{
TRUE if seriesdb matches format assumptions. Stops program if not.
}
\description{
Helper function checks whether seriesdb input corresponds to format produced by createSeriesDB.
}
\keyword{internal}
