% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{customBindRows}
\alias{customBindRows}
\title{customBindRows}
\usage{
customBindRows(i, x, y)
}
\arguments{
\item{i}{Row index as provided by a lapply call of a index list.}

\item{x}{Data frame / tibble of peaks (existing chain).}

\item{y}{Data frame / tibble of peaks (chain candidates to be added)}
}
\value{
Row bound version of x and a row of y.
}
\description{
Helper function that allows vectorized uses of bindRows over lists of chain candidates to be combined together. Each chain candidate row i from table y is merged with an existing table x. Candidates are rows of y, the index i is given by a lapply call. The result of repeated uses of this function is a list of tables x, with each x being rowbound to a row of y.
}
\details{
Context: Function is used in initializeChains.
}
\keyword{internal}
