% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportAsFeatureXML.R
\name{exportAsFeatureXML}
\alias{exportAsFeatureXML}
\title{Title}
\usage{
exportAsFeatureXML(
  peaks,
  fileTo,
  cid = "Num",
  crt = "rt",
  cmz = "mz",
  crtmin = "rtmin",
  crtmax = "rtmax",
  cmzmin = "mzmin",
  cmzmax = "mzmax",
  cquality = "sn"
)
}
\arguments{
\item{cquality}{}
}
\value{

}
\description{
Title
}
