% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{extractLongestChain}
\alias{extractLongestChain}
\title{extractLongestChain}
\usage{
extractLongestChain(chain_list)
}
\arguments{
\item{chain_list}{A list of data frames or tibbles, each representing a potential homologue series.}
}
\value{
A data frame or tibble representing a potential homologue series.
}
\description{
Helper function that extracts the data frame (data frame or tibble) with the largest number of rows from a list of data frames.
In case of a draw the first indexed chain with maximum number of rows is extracted.
}
\details{
Context: This function helps resolve multiple chain recursions by selecting the longest running series of peaks.
}
\keyword{internal}
