% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLongestMatchIndices.R
\name{extractLongestMatchIndices}
\alias{extractLongestMatchIndices}
\title{extractLongestMatchIndices}
\usage{
extractLongestMatchIndices(mz_vals, min_match_length, ppm_tolerance, step_size)
}
\arguments{
\item{mz_vals}{Vector of mz values.}

\item{min_match_length}{Minimum match length.}

\item{ppm_tolerance}{Tolerance in ppm for step size between any two mz vals.}

\item{step_size}{Mz step size to evaluate diff(mz_vals) against.}
}
\value{
Indices of mz_vals for which meet the longest subsequence criterion. If none, a boolean FALSE is returned.
}
\description{
Helper Functions evaluates a series of mz values with respect to their repetitive increment. Returns the longest sequence of mz values that follow each other within tolerance of the step size. If two sequences of mz values matching the stepsize criterion are of equal length, the first one is returned.
}
\keyword{internal}
