% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClosestMatch.R
\name{findClosestMatch}
\alias{findClosestMatch}
\title{findClosestMatch}
\usage{
findClosestMatch(
  target_mz,
  target_rt,
  sdb_row_id,
  ppm_tolerance,
  rt_tolerance,
  peak_table
)
}
\arguments{
\item{target_mz}{Target mz value around which to search.}

\item{target_rt}{Target rt value around which to search.}

\item{sdb_row_id}{Seriesdb row identifier.}

\item{ppm_tolerance}{MZ tolerance for matching in ppm.}

\item{rt_tolerance}{Retention time tolerance for matching.}

\item{peak_table}{Peak table within which to search.}
}
\value{
A tibble with 1 row and three columns; peak table row id, sdb row id and mz value.
}
\description{
Helper function that finds closest match within rt and ppm tolerance of some target mz and rt tuple in a peak_table. Extracts the row_id and mz value for further processing.
}
\keyword{internal}
