% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMzTolerance.R
\name{getMzTolerance}
\alias{getMzTolerance}
\title{getMzTolerance}
\usage{
getMzTolerance(mz, ppm_tolerance)
}
\arguments{
\item{mz}{MZ value for which the tolerance margin is to be computed.}

\item{ppm_tolerance}{Tolerance in ppm.}
}
\value{
Absolute tolerance in mz.
}
\description{
Helper function that computes the absolute mz tolerance given a mz and tolerance in ppm.
}
\keyword{internal}
