% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptbPlotInteractive.R
\name{ggplotAllHomologuesPTB}
\alias{ggplotAllHomologuesPTB}
\title{ggplotAllHomologuesPTB}
\usage{
ggplotAllHomologuesPTB(peak_table, x, y, legend_setting)
}
\arguments{
\item{peak_table}{Series database as create by detectHomologues()}

\item{x}{Range vector for xlims.}

\item{y}{Range vector for ylims.}

\item{legend_setting}{Legend setting, defaults to "none" to avoid crowded legends.}
}
\value{
A ggplot object.
}
\description{
Helper function that creates the ggplot for all homologue series in the peak_table.
}
\keyword{Internal}
