% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbPlotInteractive.R
\name{ggplotAllHomologuesSDB}
\alias{ggplotAllHomologuesSDB}
\title{ggplotAllHomologuesSDB}
\usage{
ggplotAllHomologuesSDB(seriesdb, x, y, legend_setting)
}
\arguments{
\item{seriesdb}{Series database as create by sdbCreate()}

\item{x}{Range vector for xlims.}

\item{y}{Range vector for ylims.}

\item{legend_setting}{Legend setting, defaults to "none" to avoid crowded legends.}
}
\value{
A ggplot object.
}
\description{
Helper function that creates the ggplot for all homologue series in the seriesdb.
}
\keyword{Internal}
