% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptbPlotInteractive.R
\name{ggplotFocusHomologuePTB}
\alias{ggplotFocusHomologuePTB}
\title{ggplotFocusHomologuePTB}
\usage{
ggplotFocusHomologuePTB(
  peak_table,
  selected_homologue_num,
  x,
  y,
  legend_setting
)
}
\arguments{
\item{peak_table}{Series database as create by detectHomologues()}

\item{selected_homologue_num}{homologue_id of the homologue to be highlighted.}

\item{x}{Range vector for xlims.}

\item{y}{Range vector for ylims.}

\item{legend_setting}{Legend setting, defaults to "none" to avoid crowded legends.}
}
\value{
A ggplot object.
}
\description{
Helper function that generates a focus plot for a specific homologue series.
}
\keyword{Internal}
