% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbPlotInteractive.R
\name{ggplotFocusHomologueSDB}
\alias{ggplotFocusHomologueSDB}
\title{ggplotFocusHomologuesSDB}
\usage{
ggplotFocusHomologueSDB(seriesdb, selected_homologue_num, x, y, legend_setting)
}
\arguments{
\item{seriesdb}{Series database as create by sdbCreate()}

\item{selected_homologue_num}{homologue_id of the homologue to be highlighted.}

\item{x}{Range vector for xlims.}

\item{y}{Range vector for ylims.}

\item{legend_setting}{Legend setting, defaults to "none" to avoid crowded legends.}
}
\value{
A ggplot object.
}
\description{
Helper function that creates the focus homologue plot in sdbPlotInteractive.
}
\keyword{Internal}
