% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{initializeChains}
\alias{initializeChains}
\title{initializeChains}
\usage{
initializeChains(
  peak_table,
  wmz_min,
  wmz_max,
  wrt_min,
  wrt_max,
  tolerance,
  mode,
  target_increments,
  mode2
)
}
\arguments{
\item{peak_table}{A matrix version of the peak table with mz (double), rt (double), and peak_id (int) columns.}

\item{wmz_min}{Minimum mz increment for a new peak to be considered (untargeted, absolute value)}

\item{wmz_max}{Maximum mz increment for a new peak to be considered (absolute value, untargeted)}

\item{wrt_min}{Minimum rt increment for a new peak to be considered.}

\item{wrt_max}{Maximum rt increment for a new peak to be considered.}

\item{tolerance}{Error tolerance in ppm (error relative to last chain node mz + mz stepsize).}

\item{mode}{"targeted" or "untargeted" determining the search approach.}

\item{target_increments}{Vector of increments to be searched around (targeted)}

\item{mode2}{"increment" or "decrement" to indicate whether mz steps are to be looked for in an incremental or decremental way.}
}
\value{
list of data frames containing each two rows that are a possible series start. Empty list if no candidate matches.
}
\description{
Function creates initial chain tables for the current root node (first row of peak_table). Two modes are available: "targeted" or "untargeted". In both cases, each root may have more than one possible candidate second element. Hence, a list of possible chains is returned to be followed in subsequent recursive calls of expandChains.
}
\keyword{internal}
