% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{initializeTargeted}
\alias{initializeTargeted}
\title{initializeTargeted}
\usage{
initializeTargeted(
  peak_table,
  wrt_min,
  wrt_max,
  tolerance,
  target_increments,
  mode2
)
}
\arguments{
\item{peak_table}{A matrix version of the peak table with mz (double), rt (double), and peak_id (int) columns.}

\item{wrt_min}{Minimum rt increment for a new peak to be considered.}

\item{wrt_max}{Maximum rt increment for a new peak to be considered.}

\item{tolerance}{Error tolerance in ppm (error relative to last chain node mz + mz stepsize).}

\item{target_increments}{Vector of increments to be searched around (targeted)}

\item{mode2}{"increment" or "decrement" to indicate whether mz steps are to be looked for in an incremental or decremental way.}
}
\value{
list of data frames containing each two rows that are a possible series start. Empty list if no candidate matches.
}
\description{
Function creates a (likely small) number of candidate series tables to be explored in expandChains().
}
\details{
The targeted approach creates a very narrow window using the target increments and ppm tolerances, as well as a rt tolerance to find candidate peaks to form seconds steps with respect to the root peak. Creates an initial chain for each pairwise combination of root peak + any node contained within these narrow windows.
}
\keyword{internal}
