% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{initializeUntargeted}
\alias{initializeUntargeted}
\title{initializeUntargeted}
\usage{
initializeUntargeted(
  peak_table,
  wmz_min,
  wmz_max,
  wrt_min,
  wrt_max,
  tolerance,
  mode2
)
}
\arguments{
\item{peak_table}{A matrix version of the peak table with mz (double), rt (double), and peak_id (int) columns.}

\item{wmz_min}{Minimum mz increment for a new peak to be considered (untargeted, absolute value)}

\item{wmz_max}{Maximum mz increment for a new peak to be considered (absolute value, untargeted)}

\item{wrt_min}{Minimum rt increment for a new peak to be considered.}

\item{wrt_max}{Maximum rt increment for a new peak to be considered.}

\item{tolerance}{Error tolerance in ppm (error relative to last chain node mz + mz stepsize).}

\item{mode2}{"increment" or "decrement" to indicate whether mz steps are to be looked for in an incremental or decremental way.}
}
\value{
list of data frames containing each two rows that are a possible series start. Empty list if no candidate matches.
}
\description{
Function creates a (possibly large) number of candidate series tables to be explored in expandChains(). The approach creates a filter window using mz and rt bounds, and creates an initial chain for each pairwise combination of root + any node in these windows.
}
\keyword{internal}
