% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isScalar}
\alias{isScalar}
\title{isScalar}
\usage{
isScalar(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Boolean True or False.
}
\description{
Function checks whether input is atomic, and if so, of length 1.
}
\keyword{internal}
