% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtbls1358}
\alias{mtbls1358}
\title{MTBLS1358 dataset}
\format{
A data frame with 16,106 rows and 4 variables:
\describe{
\item{peak_id}{unique identifier for each feature, type integer}
\item{mz}{mass-to-charge ratio, type double}
\item{rt}{retention time in seconds, type double}
\item{intensity}{average feature intensity across the different samples of the original MTBLS1358 dataset, type double}
}
}
\source{

}
\usage{
mtbls1358
}
\description{
This dataset is used as an example peak table in the vignette of homologueDiscoverer. Peak table with peak_id, mz, rt and intensity columns from the MTBLS1358 dataset.
}
\keyword{datasets}
