% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{nearSelect}
\alias{nearSelect}
\title{nearSelect}
\usage{
nearSelect(value, table, tolerance)
}
\arguments{
\item{value}{The target mz value for searching.}

\item{table}{Peak table in matrix form, with named column "mz".}

\item{tolerance}{Relative error tolerance specified in ppm.}
}
\value{
Table filtered down to those rows matching the search criteria. May be empty, nrow = 0, but remains a data frame type object.
}
\description{
Helper function filters down a peak table (matrix format) to contain only mz rows falling within upper bounds and lower bounds determined via the target mz value and an error tolerance in ppm.
}
\details{
Context: is used in vectorized nearSelect. Used in initializeChains.
}
\keyword{internal}
