% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series_db.R
\name{plotSeriesDBInteractive}
\alias{plotSeriesDBInteractive}
\title{plotPeakTableInteractive}
\usage{
plotSeriesDBInteractive(seriesdb, max_legend_size = 100)
}
\arguments{
\item{seriesdb}{}
}
\description{
Opens s shiny web-application with an interactive plot depicting all the series in the provided series data base. Accepts the series
}
