% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotateHomologues.R
\name{preScreenPeakTable}
\alias{preScreenPeakTable}
\title{preScreenPeakTable}
\usage{
preScreenPeakTable(peak_table, seriesdb, ppm_tolerance, rt_tolerance)
}
\arguments{
\item{peak_table}{A peak table with mz and rt columnds.}

\item{seriesdb}{A seriesdb tibble created with createSeriesDB().}

\item{ppm_tolerance}{Tolerance in mz for matching two peaks.}

\item{rt_tolerance}{Retention time tolerance in matching. Beware of rt units in the database!}
}
\description{
Helper function of annotateHomologues that checks whether seriesdb max intensity peaks of the various homologues have a match within ppm and rt tolerance in the peak table. Filters seriesdb down to those homologue ids which do have a match for complete series matching in annotateHomologues.
}
\keyword{internal}
