% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_detectHomologues.R
\name{printProgress}
\alias{printProgress}
\title{printProgress}
\usage{
printProgress(peak_table, progress_counter)
}
\arguments{
\item{peak_table}{Peak table in matrix form.}

\item{progress_counter}{Integer value determining progress print. If nrow peak table is a multiple of progress counter, the remaining number of peaks is printed.}
}
\description{
Helper function which prints progress dependent on the number of rows left in peak table (matrix) and a progress counter indicting the message frequency.
}
\keyword{internal}
