% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptbPlotInteractive.R
\name{ptbPlotInteractive}
\alias{ptbPlotInteractive}
\title{ptbPlotInteractive}
\usage{
ptbPlotInteractive(peak_table, legend_setting = "none", plot_height = 600)
}
\arguments{
\item{peak_table}{A annotated peak table as produced by detectHomologues()}

\item{legend_setting}{A legend setting (position i.e. "bottom", "left", or default = "none")}

\item{plot_height}{Plot height in pixels. Defaults to 600. Plot always adjusts width to window size.}
}
\value{
None. An interactive plot is opened.
}
\description{
Functions creates interactive R Shiny server (local) for visualizing a peak_table created using detectHomologues().
}
\details{
Figure can be interacted with in the following ways:

Left click and drag to select an area of the figure for which to show the corresponding data in the table below the figure.

Left click and drag to select an area of the plot. Use double click to zoom into the area.

Double click with no selection zooms out to original view.

Select a polymer id using the input option. Click on the toggle plot button to change the plot type and table to highlight the selected homologue series.
}
