% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbCheckContained.R
\name{sdbCheckContained}
\alias{sdbCheckContained}
\title{sdbCheckContained}
\usage{
sdbCheckContained(seriesdb, new_series, ppm_tolerance, rt_tolerance)
}
\arguments{
\item{seriesdb}{A seriesdb tibble created with createSeriesDB().}

\item{new_series}{Annotated series output generated by detectHomologues. By default, any homologue identifier equaling NA is omitted as this identifier is the default for non-series peaks.}

\item{ppm_tolerance}{Tolerance in ppm for matching two peaks.}

\item{rt_tolerance}{Retention time tolerance in matching. Beware of rt units in the database and peak table!}
}
\value{
A list of integer homologue identifiers that were found to be overlapping between new_series and seriesdb.
}
\description{
Function checks for peak overlap between series in new_series tibble (annotated peak table) and series_db tibbles. If any peaks of a homologue_id in new_series is found in series_db, it's id is returned to prevent peak overlaps between series in seriesdb.
}
\details{
Please note that peak matching is fuzzy, and depending on settings two peaks may or may not be considered overlapping. Matching stringency should be held similar over sdbCheckContained and annotateHomologues to avoid annotation overwriting caused by peak overlaps.
}
