% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbCreate.R
\name{sdbCreate}
\alias{sdbCreate}
\title{sdbCreate}
\usage{
sdbCreate(
  annotated_peak_table,
  annotation_id = NA,
  noise = TRUE,
  homologue_series_name = NA,
  sample_origin = NA,
  sample_description = NA
)
}
\arguments{
\item{annotated_peak_table}{An annotated peak table in tibble format as produced by identifyHomologues.}

\item{annotation_id}{Integer scalar or vector of ids to include in the seriesdb.}

\item{noise}{Boolean scalar or vector of same length as annotation id with noise status for homologue_ids in annotation_id.}

\item{homologue_series_name}{Character scalar or vector of same length as annotation id with names for homologue_ids in annotation_id.}

\item{sample_origin}{Character string indicating the sample / processing run origin of the identified series in non-standardized.}

\item{sample_description}{Character string containing any relevant information about the sample / processing run in non-standardized from.}
}
\value{
A series_db tibble.
}
\description{
Function creates series_db object from annotated_peak_table. Scalar or vector inputs for annotation_id (integer), noise (boolean) and homologue_series_name (character) can be provided to select specific homologue_ids from the annotated peak_table and give them a manual noise status or homologue series name. By default all three inputs are NA, leading to all homologue_ids from the annotated peak table being used, with noise = TRUE and homologue_series_name = NA.
}
