% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_functions.R
\name{sdbLoad}
\alias{sdbLoad}
\title{sdbLoad}
\usage{
sdbLoad(series_db_path)
}
\arguments{
\item{series_db_path}{Absolute or relative path from current working directory to a saved series_db file.}
}
\value{
series_db tibble.
}
\description{
Function loads a series_db object from provided path. A series_db format check is performed before returning the read object to alert users to possible mismatch in data structure. Function stopped if wrong format detected. series_db object needs to be saved as .RDS file using sdbSave or saveRDS.
}
