% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_functions.R
\name{sdbPush}
\alias{sdbPush}
\title{sdbPush}
\usage{
sdbPush(series_db, new_series, sample_origin = NA, sample_description = NA)
}
\arguments{
\item{series_db}{A series_db tibble as generated by sdbCreate().}

\item{new_series}{A annotated peak_table as produced by detectHomologues().}

\item{sample_origin}{Character string indicating the sample / processing run origin of the identified series in non-standardized form.}

\item{sample_description}{Character string containing any relevant information about the sample / processing run in non-standardized from.}
}
\value{
series_db A series_db tibble as generated by sdbCreate() with an additional set of homologues added.
}
\description{
Function adds homologues from new_series tibble (annotated peak table) to series_db. All new additions will receive their unique push_id and timestamp. No duplicate checking is done here, please use checkContained to make sure that homologues or their constituent peaks to be added are not already contained within series_db.
}
