% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_functions.R
\name{sdbRenameHomologue}
\alias{sdbRenameHomologue}
\title{sdbRenameHomologue}
\usage{
sdbRenameHomologue(series_db, homologue_id_set, name_set)
}
\arguments{
\item{series_db}{A series_db tibble as generated by sdbCreate()}

\item{homologue_id_set}{An integer scalar or list/vector of integer homologue ids to be renamed.}

\item{name_set}{A character string or list/vector of character strings with names for the homologue_id_set.}
}
\value{
A series_db tibble with updated homologue names.
}
\description{
Function allows for renaming specific homologue_ids for future reference.
}
