% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_functions.R
\name{sdbRescaleHomologueIds}
\alias{sdbRescaleHomologueIds}
\title{sdbRescaleHomologueIds}
\usage{
sdbRescaleHomologueIds(series_db)
}
\arguments{
\item{series_db}{A series_db tibble as generated by sdbCreate()}
}
\value{
A series_db tibble with homologue ids rescaled to range from 1 to n_homologues.
}
\description{
Function changes homologue_ids in series_db to consecutive integers starting from 0 to the number of homologues in series_db. homologue_id ordering is maintained. Utility function designed to have simpler homologue identifiers for plotting if many homologue ids are removed from series_db. Homologue names and noise status remain unchanged.
}
\details{
Changing homologue_ids will lead to identifier inconsistencies with any analyses or plots created prior to renaming!
}
