% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management_functions.R
\name{sdbSetHomologueNoise}
\alias{sdbSetHomologueNoise}
\title{sdbSetHomologueNoise}
\usage{
sdbSetHomologueNoise(series_db, homologue_id_set, noise_set)
}
\arguments{
\item{series_db}{A series_db tibble as generated by sdbCreate()}

\item{homologue_id_set}{An integer scalar or list/vector of integer homologue ids to have their noise status changed.}

\item{noise_set}{An boolean scalar or list/vector of booleans with noise assignments for the homologue_id_set.}
}
\value{
A series_db tibble with updated homologue noise identifiers.
}
\description{
Function allows to change the noise attribute of a single or set of homologue ids within series_db.
}
\details{
If a set of homologue_ids are to get the same noise level please use: rep(TRUE, times = length(homologue_ids)) or rep(FALSE, times = length(homologue_ids)) to guarantee length of noise vector is equal to length of homologue_id vector.
}
