% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbSummarize.R
\name{sdbSummarize}
\alias{sdbSummarize}
\title{sdbSummarize}
\usage{
sdbSummarize(seriesdb, calcMF_tolerance = 100)
}
\arguments{
\item{seriesdb}{A seriesdb tibble created with createSeriesDB().}

\item{calcMF_tolerance}{Maximum ppm tolerance for MassTools::CalcMF.}
}
\value{
Summary table in tibble format.
}
\description{
Function generates a summary table with information about the different homologues in seriesdb. The summary table is printed with 10 decimals as well as returned as a tibble.
}
\details{
For each homologue id in seriesdb, the summary contains the series_length (number of peaks in series), the mean and median differences in mz between peaks, as well as the molecular formula estimate and the theoretical increment estimate as computed by MassTools::CalcMF.
}
