% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdbSummarize.R
\name{wrapperCalcMF}
\alias{wrapperCalcMF}
\title{wrapperCalcMF}
\usage{
wrapperCalcMF(mz, tolerance)
}
\arguments{
\item{mz}{Mass to charge value for which to calculate Molecular Formula and Theoretical Mass.}

\item{tolerance}{Maximum ppm tolerance for MassTools::CalcMF.}
}
\value{
Tibble with one row and two columns for molecular_formula (character) and theoretical_increment (double).
}
\description{
Helper function for sdbSummarize that uses MassTools::calcMF on a provided mz value with given tolerance, and extracts the hit with the lowest ppm, if any. Results are returned as a tibble with one row and a molecular formula as well as a theoretical increment.
}
\keyword{internal}
